requirements.txt
accelerate
annotated-types
anyio
appdirs
asgiref
attrs
autobahn
Automat
black
cachetools
certifi
cffi
channels
channels_redis
charset-normalizer
click
colorama
constantly
cryptography
daphne
defusedxml
Django
django-autoslug
django-cors-headers
django-countries
django-environ
django-filter
django-phonenumber-field
djangorestframework
djangorestframework_simplejwt
djoser
environ
filelock
flake8
fsspec
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
huggingface-hub
hyperlink
idna
incremental
isort
Jinja2
MarkupSafe
mccabe
mpmath
msgpack
mypy-extensions
networkx
numpy
oauthlib
packaging
pathspec
phonenumbers
pillow
proto-plus
protobuf
psutil
psycopg2-binary
pyasn1
pyasn1_modules
pycodestyle
pycparser
pydantic
pydantic_core
pyflakes
PyJWT
pyOpenSSL
pyparsing
python3-openid
pytz
PyYAML
redis
regex
requests
requests-oauthlib
rsa
safetensors
service-identity
setuptools
sniffio
social-auth-app-django
social-auth-core
sqlparse
sympy
tokenizers
toml
torch
tqdm
transformers
Twisted
txaio
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
websockets
zope.interface
